LayoutTests/platform/wpe/imported/w3c/web-platform-tests/encrypted-media/clearkey-keystatuses-multiple-sessions.https-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: undefined is not an object (evaluating 'contentitem.keys')

Harness Error (FAIL), message = Unhandled rejection: undefined is not an object (evaluating 'contentitem.keys')
